requirements.txt
pydantic~=2.9.2
gpt_researcher
git+https://github.com/Leea-Labs/leea-agent-sdk@main
python-dotenv
